v0.4.4
update on bug: im pretty sure the head isnt correctly updating. i dont know why. try 1, 2, 3. What happens is it does printTree from 1, which is the 'old' head that got rotated in the left rotation upon inserting 3. even if i manually set the head pointer to point at any parent of head (which should be 2 in this case) AT THE END of the rotation method, once outside and i call printTree, it just prints 1. So the head outside the method is different from the head inside...